move multipart code into own file
